Potential fix for the font issue
After some testing, it looks like the only way to remove the poppins cors issue and not have to use the production url for the fonts folder is to host all the fonts in the themes/fonts folder.

I'm also testing the performance of the About page where the div with the background does not have inline styles and the mobile image is referenced on the page.